Moved some components for better organization
