+++ socialtours/src/App.js
@@ -18,8 +18,8 @@ import Protected from "./components/Protected";
 import NotFound from "./components/NotFound";
 import Callback from "./components/UserComponents/Callback";
 import Calendar from "./components/EventCalendar";
-import UpdateEvent from "./components/UpdateEvent.js";
-import TheCreateEvent from "./components/createEvent.js";
+import UpdateEvent from "./components/ManageEvents/UpdateEvent.js";
+import TheCreateEvent from "./components/ManageEvents/createEvent.js";
 //import TheCrudEvent from "./components/updateDeleteEvent.js";
 import ManageEvents from "./components/ManageEvents/ManageEvents";
 import Navigation from "./components/Navigation";
