+++ socialtours/src/components/ManageEvents/ScheduleEvent.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ScheduleEvent = props => {
+	return <div>Schedule event</div>;
+};
+
+export default ScheduleEvent;
